Added fast batch2d

+++ src/window/glfw_window.cpp
@@ -111,7 +111,7 @@ void glfw_window::update()
     glfwPollEvents();
 }
 
-float glfw_window::getTicks()
+double glfw_window::getTicks()
 {
     return glfwGetTime();
 }
